Move cloud under pkg/cloud
