+++ pkg/cloud/libvirt/actuators/machine/utils/ignition.go
