+++ pkg/cloud/libvirt/actuators/machine/utils/image.go
