+++ pkg/cloud/libvirt/actuators/machine/utils/network.go
