+++ pkg/cloud/libvirt/actuators/machine/utils/stream.go
