+++ pkg/cloud/libvirt/providerconfig/v1alpha1/doc.go
@@ -12,7 +12,7 @@
 // limitations under the License.
 
 // +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=github.com/openshift/cluster-api-provider-libvirt/cloud/libvirt/providerconfig
+// +k8s:conversion-gen=github.com/openshift/cluster-api-provider-libvirt/pkg/cloud/libvirt/providerconfig
 // +k8s:openapi-gen=true
 // +k8s:defaulter-gen=TypeMeta
