+++ pkg/cloud/libvirt/providerconfig/v1alpha1/zz_generated.deepcopy.go
